docs: improve the clarity of the documentation
